+++ CHANGELOG.md
@@ -15,10 +15,12 @@ which adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ### Added:
 
-- License added.
+- License.
 - Standard project structure for the use of this extension.
 - Support for more than one application/sub-program.
 - Support for more than one entry point per application/sub-program.
+- VSC Settings for project creation and initialization.
+- C/C++ settings for project creation and initialization.
 - `ESP32-PM: Register ESP-IDF API` command for Windows.
 - `ESP32-PM: Register Espressif Toolchain` command for Windows.
 - `ESP32-PM: Create project` command for Windows.