+++ CHANGELOG.md
@@ -5,9 +5,15 @@ All notable changes to the "ESP32 project manager" extension will be documented 
 **Note:** The format of this changelog is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 which adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+<!-- TODO #2 Replace the [Unreleased] tag by the new version number -->
 ## [Unreleased]
 
+<!-- TODO #3 Update the repo link -->
 [Repository](https://github.com/mrverdant13/esp32-pm-vsc-extension/tree/develop)
+
+## [0.1.0] - 2019-12-02
+
+[Repository](https://github.com/mrverdant13/esp32-pm-vsc-extension/tree/v0.1.0)
 
 ### Added:
 
@@ -28,3 +34,7 @@ which adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `ESP32-PM: Monitor` command for Windows and Linux.
 - `ESP32-PM: Flash & Monitor` command for Windows and Linux.
 - `ESP32-PM: Clean` command for Windows and Linux.
+
+# Comparison
+
+- [v0.1.0](https://github.com/mrverdant13/esp32-pm-vsc-extension/releases/tag/v0.0.1)